Data-Driven Methods
These files are for generating data iteratively, to compare the performance of different algorithms with random network  
configurations. When used in conjunction to the random graph generator, you can benchmark any of the algorithms with an N sized networks to see which one is best. 

Brute force algorithms are not ideal for larger network sizes, and usually take a long time unless you multiprocess the entire search, but then it becomes quite energy intensive.